rearrange tests to follow library structure
